Fixed: Fujian works with CPython 2.7
Fujian didn't work with CPython 2.7 because the __builtins__ module could not
be deepcopied. Now, Fujian shares its __builtins__ module with the code being
executed, so the deepcopy is unnecessary, so Fujian works with CPython 2.7.

This doesn create another avenue for submitted code to crash Fujian, but if
someone starts reassigning things in __builtins__ then they're really asking
for trouble to begin with.